+++ src/routes/sub-layouts/(first)/+layout.server.ts
@@ -1,0 +1,8 @@
+/*
+import type { LayoutServerLoad } from "./$types";
+
+export const load = (async ({ parent }) => {
+  const { flash } = await parent();
+  return { flash };
+}) satisfies LayoutServerLoad;
+*/