+++ src/routes/sub-layouts/(first)/+page.server.ts
@@ -1,0 +1,7 @@
+import { redirect } from '$lib/server';
+
+export const actions = {
+  default: ({ cookies }) => {
+    redirect(303, '/sub-layouts/route2', { message: 'Redirected to route2' }, cookies);
+  }
+};