+++ CHANGELOG.md
@@ -5,6 +5,12 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.4.6] - 2025-01-23
+
+### Fixed
+
+- Possible navigation regression fix for Svelte 5.
+
 ## [2.4.5] - 2025-01-07
 
 ### Fixed